services/worker-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth==2.25.2
requests==2.31.0
redis==5.0.1
celery==5.3.4
